fix integer wrapping and new PIDs

+++ GimbalFirmware/Headers/control/gyro_kinematics_correction.h
@@ -11,6 +11,6 @@ extern float sin_theta;
 extern float cos_theta;
 
 void update_joint_ang_trig();
-void transform_ang_vel_to_joint_rate(int16* gyro_in, int16* gyro_out);
+void transform_ang_vel_to_joint_rate(float* gyro_in, float* gyro_out);
 
 #endif /* GYRO_KINEMATICS_CORRECTION_H_ */
